requirements.dev.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pytest
httpx
grpcio==1.65.1
numpy==1.26.4
PyJWT==2.8.0
protobuf==4.25.3
requests==2.32.3
tokenizers==0.19.1
transformers==4.42.4
yandexcloud==0.302.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.21
langchain-text-splitters==0.2.2
scikit-learn==1.5.1
pymongo
motor==3.5.1
